==> jupyter/08_phenotype_wrangling.ipynb <==
import pandas as pd
import numpy as np

src = '../hpcc/traditional/'
filename = '../all_phenos_for_Erik_221024.csv'
attr = pd.read_csv(filename)
print(attr.shape)

col_names = ['bname', 'wname',
             'nut_length','nut_height', 'nut_width',
             'nut_vol',
             'nut_va3d',
             'nut_ratio_feret',
             'nut_area',
             'nut_sphericity',
             'nut_cnvx_area',
             'nut_cnvx_vol',
             'nut_ratio_cnvx_area',
             'nut_inv_ratio_cnvx_vol',
             'krumbein',
             'sneed',
             'air_vol', 'meat_vol', 'shell_vol', 'vein_vol',
             'air_ratio_vol', 'meat_ratio_vol', 'shell_ratio_vol', 'vein_ratio_vol',
             'shell_rugosity',
             'shell_thickness',
             'shell_ext_ratio', 'shell_prot_ratio',
             'shell_ext_vol', 'shell_prot_vol',
             'meat_length', 'meat_height', 'meat_width',
             'meat_area',
             'meat_cnvx_vol',
             'meat_cnvx_area',
             'meat_ratio_cnvx_area',
             'meat_inv_ratio_cnvx_vol',
             'meat_cavity_area',
             'meat_cavity_vol',
             'meat_arc_height',
             'meat_cavity_height',
             'meat_sphericity',
             'meat_ratio_cavity_area',
             'meat_ratio_cavity_vol',
             'rho_mvs', 'rho_vvs', 'rho_vvm'
             ]

filename = src + 'traditional_raw.csv'
raw = pd.read_csv(filename, header=None, dtype={1:str}, names=col_names)
raw.loc[raw['bname'] == '201SBa_R5_T55', 'bname'] = '2014SBa_R5_T55'
foo = raw.iloc[:, 30:33].values
MM = np.max(foo, axis=1)
mm = np.min(foo, axis=1)
raw['meat_ratio_feret'] = MM/mm

raw['nut_vol_sa'] = raw['meat_vol'] + raw['vein_vol'] + raw['shell_vol']

raw['meat_ratio_vol_sa'] = raw['meat_vol']/raw['nut_vol_sa']
raw['vein_ratio_vol_sa'] = raw['vein_vol']/raw['nut_vol_sa']
raw['shell_ratio_vol_sa'] = raw['shell_vol']/raw['nut_vol_sa']

raw['meat_ratio_cnvx_area'] = 1/raw['meat_ratio_cnvx_area']
raw['meat_inv_ratio_cnvx_vol'] = 1/raw['meat_inv_ratio_cnvx_vol']
raw = raw.drop(columns=['shell_ext_ratio', 'shell_ext_vol'])

#data = data.drop(columns = [''])
print(raw.shape)
raw.head()

cattraits = np.zeros((len(raw), 15)) - 1
UCACCSD = []

for i in range(len(cattraits)):
    locate, Rrow, Ttree = raw.iloc[i,0].split('_')
    row = int(Rrow[1:])
    tree = int(Ttree[1:])

    if locate == 'NewStuke':
        locate = 'New Stuke'
    elif locate in ['SelectD', 'SelectF', 'SelectC']:
        locate = 'Select Davis'

    metadf = attr[(attr['LOCATE']==locate) & (attr['ROW']==row) & (attr['TREE']==tree)]
    if len(metadf) > 0:
        ucaccsd = metadf['UCACCSD'].iloc[0]
    else:
        ucaccsd = ''
        print(raw.iloc[i,0])
    
    UCACCSD.append(ucaccsd)

raw['UCACCSD'] = UCACCSD
foo = np.asarray([51] + list(range(51)))
raw = raw.iloc[:, foo]
print(raw.shape)
raw.head()

b = '2008SB_R4_T13'
b.split('_')

bname = []
lname = []
rname = []
tname = []
uname = []
N = []
M = []

for b in np.sort(pd.unique(raw.bname)):
    l,r,t = b.split('_')
    
    if l == 'NewStuke':
        l = 'New Stuke'
    elif l in ['SelectD', 'SelectF', 'SelectC']:
        l = 'Select Davis'
        
    bname.append(b)
    lname.append(l)
    rname.append(r[1:])
    tname.append(t[1:])
    foo = raw[raw.bname == b]
    uname.append(foo['UCACCSD'].iloc[0])
    N.append(len(foo))
    M.append(np.sum(foo.meat_ratio_vol > floor))

rawdf = pd.DataFrame()
rawdf['Scan Number'] = np.arange(1,len(bname) + 1)
rawdf['Scan Name'] = bname
rawdf['Locate'] = lname
rawdf['Row Number'] = rname
rawdf['Tree Number'] = tname
rawdf['UCACCSD'] = uname
rawdf['Number of nuts scanned'] = N
rawdf['Number of nuts preserved'] = M
print(rawdf.shape)
rawdf.head()
rawdf.to_csv('../S1_scanned_batches.csv', index=False)

kratio = raw.loc[:,'meat_ratio_vol'].values
q25,q50,q75 = np.quantile(kratio, [0.25,.5,0.75])
print(q25,q50,q75, sep='\t')
iqr = q75 - q25
floor = q25-2*iqr
print(floor)
nonemptyk = np.where(kratio > floor)[0]

data = raw.iloc[nonemptyk, :]
print(data.shape)

filename = src + 'traditional_phenotype_summary.csv'
print(filename)
#data.to_csv(filename, index=False)

data.head()

data.meat_ratio_vol_sa - data.meat_ratio_vol/(1-data.air_ratio_vol)

attr[attr['EaseOfRemoval']==3]

data[data['UCACCSD'] == '12-045-8']

# Per accession averaging

ucaccsd = sorted(np.unique(UCACCSD))[1:]
print(len(ucaccsd))

ucaccsd = sorted(np.unique(data['UCACCSD']))[1:]
print(len(ucaccsd))

quant = np.zeros((len(ucaccsd), data.shape[1]-3 + attr.shape[1] - 5 + 2))

for i in range(len(quant)):
    uc = ucaccsd[i]
    
    foo = data[data['UCACCSD'] == uc]
    bar = attr[attr['UCACCSD'] == uc]
    
    quant[i] = np.hstack(([len(foo), len(bar)], 
                          np.nanmean(foo.iloc[:,iniN:].values, axis = 0),
                          np.nanmean(bar.iloc[:,5:].values, axis = 0)))

uc = '12-045-8'
uc = '91-136'
#uc = '85-023-2'
foo = data[data['UCACCSD'] == uc]
bar = attr[attr['UCACCSD'] == uc]

bar.iloc[:,5:].describe().loc[['mean','std']]

bar

col = ['nut_count', 'plot_count'] + list(data.columns[3:]) + list(attr.columns[5:])
qual = pd.DataFrame(quant, columns = col)
qual = qual.astype({'nut_count': int, 'plot_count':int})
foo = np.asarray([qual.shape[1]] + list(range(qual.shape[1])))

qual['UCACCSD'] = ucaccsd

qual = qual.iloc[:, foo]
print(qual.shape)

filename = src + 'qual_quant_accession_summary.csv'
print(filename)
qual.to_csv(filename, index=False)

qual.head()

np.min(qual['EaseOfRemoval'])

# Per tree averaging

ucaccsd = sorted(np.unique(data['bname']))[1:]
print(len(ucaccsd))

i = 0
uc = ucaccsd[i]
print(uc)
locale, row, tree = uc.split('_')
row = int(row[1:])
tree = int(tree[1:])
print(locale, row, tree)
attr[(attr.LOCATE == locale) & (attr.ROW == row) & (attr.TREE == tree)]

idu = []
quant = np.zeros((len(ucaccsd), data.shape[1]-3 + attr.shape[1] - 5 + 2))
print(quant.shape)
i = 0
for i in range(len(quant)):
    uc = ucaccsd[i]
    locale, row, tree = uc.split('_')
    if locale == 'NewStuke':
        locale = 'New Stuke'
    elif locale in ['SelectC','SelectD','SelectF']:
        locale = 'Select Davis'
    row = int(row[1:])
    tree = int(tree[1:])
    
    foo = data[data['bname'] == uc]
    idu.append(foo.UCACCSD.iloc[0])
    bar = attr[(attr.LOCATE == locale) & (attr.ROW == row) & (attr.TREE == tree)]
    if len(bar) > 0:
    
        quant[i] = np.hstack(([len(foo), len(bar)], 
                          np.nanmean(foo.iloc[:,3:].values, axis = 0),
                          np.nanmean(bar.iloc[:,5:].values, axis = 0)))

col = ['nut_count', 'tree_count'] + list(data.columns[3:]) + list(attr.columns[5:])
qual = pd.DataFrame(quant, columns = col)
qual = qual.astype({'nut_count': int, 'tree_count':int})
foo = np.asarray([qual.shape[1]] + list(range(qual.shape[1])))
qual['UCACCSD'] = [idu[i] + '/' + ucaccsd[i] for i in range(len(idu))]

qual = qual.iloc[~np.all(quant == 0, axis=1), foo]
print(qual.shape)


filename = src + 'qual_quant_tree_summary.csv'
print(filename)
qual.to_csv(filename, index=False)

qual.head()

# Summary

col_labels = ['UCACCSD', 'Batch Name', 'Scan Name',
             'Nut Length','Nut Height', 'Nut Width',
             'Nut Volume',
             'Nut VA3D',
             'Nut Feret Ratio',
             'Nut Surface Area',
             'Nut Sphericity',
             'Nut Convex Surface Area',
             'Nut Convex Volume',
             'Nut Ratio Convex Area',
             'Nut Ratio Convex Volume',
             'Krumbein Index',
             'Sneed Index',
             'Air Volume', 'Kernel Volume', 'Shell Volume', 'Packing Volume',
             'Air Vol Ratio', 'Kernel Vol Ratio', 'Shell Vol Ratio', 'Packing Vol Ratio',
             'Shell Rugosity',
             'Shell Thickness',
             'Protruding Shell Ratio', #
             'Protruding Shell Volume', #
             'Kernel Length', 'Kernel Height', 'Kernel Width',
             'Kernel Surface Area',
             'Kernel Convex Volume',
             'Kernel Convex Area',
             'Kernel Ratio Convex Area',
             'Kernel Ratio Convex Volume',
             'Kernel Cavity Area',
             'Kernel Cavity Volume',
             'Kernel Arc Height',
             'Kernel Cavity Height',
             'Kernel Sphericity',
             'Kernel Ratio Cavity Area',
             'Kernel Ratio Cavity Volume',
             'Density Kernel vs Shell', 'Density Packing vs Shell', 'Density Packing vs Kernel',
             'Kernel Feret Ratio',
             'Nut Vol wo Air', 
             'Kernel Vol Ratio wo Air', 'Packing Vol Ratio wo Air', 'Shell Vol Ratio wo Air',
             'Shell Integrity',
             'Shell Texture',
             'Shell Color',
             'Shell Seal',
             'Shell Strength',
             'Shell Thickness',
             'Packing Tissue',
             'Kernel Fill',
             'Tip Shrivel',
             'Minor Shrivel',
             'Major Shrivel',
             'Plumpness',
             'Ease Of Removal','Kernel Weight Ratio', 'Blank'
             ]

print(len(col_labels))

col_units = ['', '', '',
             ' [mm]',' [mm]', ' [mm]',
             ' [mm$^3$]',
             '',
             '',
             ' [mm$^2$]',
             '',
             ' [mm$^2$]',
             ' [mm$^3$]',
             ' [%]',
             ' [%]',
             '',
             '',
             ' [mm$^3$]', ' [mm$^3$]', ' [mm$^3$]', ' [mm$^3$]',
             ' [%]', ' [%]', ' [%]', ' [%]',
             '',
             ' [mm]',
             ' [%]', #
             ' [mm$^3$]', #
             ' [mm]', ' [mm]', ' [mm]',
             ' [mm$^2$]',
             ' [mm$^3$]',
             ' [mm$^2$]',
             ' [%]',
             ' [%]',
             ' [mm$^2$]',
             ' [mm$^3$]',
             ' [mm]',
             ' [mm]',
             '',
             ' [%]',
             ' [%]',
             ' [%]', ' [%]', ' [%]',
             '',
             ' [mm$^3$]',
             ' [%]', '[%]', ' [%]',
             '',
             '',
             '',
             '',
             '',
             '',
             '',
             '',
             ' [%]',
             ' [%]',
             ' [%]',
             '',
             '',' [%]', ' [%]'
             ]

foo = pd.DataFrame()
foo['col_labels'] = col_labels
foo['col_units'] = col_units

filename = src + 'col_labels.csv'
foo.to_csv(filename, index=False)

for i in range(len(col_labels)):
    print(i, i-3, col_labels[i],  qual.columns[i], sep='\t')
