==> jupyter/16_ect_wrangling.ipynb <==
import numpy as np
import pandas as pd
import os
import glob 

src = '../hpcc/topology/'
dst = '../hpcc/results/ect/'

DD = 50
imgname = 'walnut'
for DD in [50,100,150,200,300,500]:
    files = sorted(glob.glob(src + '*/*_'+imgname+'_d{:04d}_*.csv'.format(DD)))
    print(len(files))
    bname = ['' for i in range(len(files))]
    lname = ['' for i in range(len(files))]

    for i in range(len(files)):
        foo = os.path.split(files[i])[1].split('_')
        bname[i] = '_'.join(foo[:3])
        lname[i] = foo[3]
    ects = np.zeros((len(files), DD*64), dtype=int)
    for i in range(len(files)):
        ects[i] = np.loadtxt(files[i], delimiter=',', dtype=int)

    ect = pd.DataFrame()
    ect['bname'] = bname
    ect['lname'] = lname
    filename = dst + 'ects_' + imgname + '_d{:04d}_t64.csv'.format(DD)
    pd.concat([ect, pd.DataFrame(ects)], ignore_index=True, axis=1).to_csv(filename, index=False)

idname = '2014SBa'

rownum = 1
treenum = 36
nutnum = 1

bname = idname + '_R{}_T{}'.format(rownum, treenum)
fname = bname + '_{:03d}'.format(nutnum)
print(fname)

esrc = src + bname + '/' + fname + '_'
DD = 150
TT = [4,8,16,32,64]

ects = [None for i in range(len(TT))]
for i in range(len(ects)):
    ects[i] = np.loadtxt(esrc + 'd{:04d}_t{:02d}.csv'.format(DD,TT[i]), delimiter=',', dtype=int)

for i in range(len(ects)):
    print(i, TT[i], len(ects[i]), sep='\t')

for i in range(len(ects)):
    print(i, ects[i][:15], sep='\n')

idx = 3
j = 4 - idx
foo = ects[4][(2**j-1)::2**j]
print(len(foo))
np.sum(np.abs(foo - ects[idx]))

Tvals = [4,8,16,32,64]
Tmax = max(Tvals)
Tmax

csrc = '../hpcc/clean/'
walnut_files = sorted(glob.glob(csrc + bname + '/*.tif'))
widx = 0
pa, fi = os.path.split(walnut_files[widx])
fname = os.path.splitext(fi)[0]
shell_file = src + bname + '/' + fname + '_shell.tif'
kernel_file = src + bname + '/' + fname + '_meat.tif'
print(fname)

for imgfile, imgname in zip([walnut_files[widx], shell_file, kernel_file], ['walnut', 'shell', 'kernel']):
    print(imgfile, imgname, sep='\t')
